refactor: MomentToolbarHandler 이름 변경
- 이전: MomentToolbarHandler
- 이후: StaccatoToolbarHandler
