+++ android/Staccato_AN/app/src/main/java/com/on/staccato/presentation/staccato/feeling/MomentFeelingSelectionFragment.kt
@@ -6,8 +6,8 @@ import androidx.fragment.app.viewModels
 import com.on.staccato.R
 import com.on.staccato.databinding.FragmentMomentFeelingSelectionBinding
 import com.on.staccato.presentation.base.BindingFragment
-import com.on.staccato.presentation.staccato.MomentFragment.Companion.DEFAULT_STACCATO_ID
-import com.on.staccato.presentation.staccato.MomentFragment.Companion.STACCATO_ID_KEY
+import com.on.staccato.presentation.staccato.StaccatoFragment.Companion.DEFAULT_STACCATO_ID
+import com.on.staccato.presentation.staccato.StaccatoFragment.Companion.STACCATO_ID_KEY
 import com.on.staccato.presentation.staccato.viewmodel.StaccatoViewModel
 import dagger.hilt.android.AndroidEntryPoint
