+++ android/Staccato_AN/app/src/main/res/layout/fragment_moment.xml
@@ -10,7 +10,7 @@
 
         <variable
             name="toolbarHandler"
-            type="com.on.staccato.presentation.staccato.MomentToolbarHandler" />
+            type="com.on.staccato.presentation.staccato.StaccatoToolbarHandler" />
 
         <variable
             name="viewModel"
